Add _thread.CriticalLock which prevents the current thread from parking
There is a deadlock hazard when a thread acquires a lock in a finalizer
that may also be held during a stop-the-world GC. Typical GC
implementations avoid this by running finalizers in separate thread
while other threads are resumed. This is awkward in Python because
some code and tests expect finalizers to have finished by the time
gc.collect() finishes.

In CPython with the GIL the issue can be mitigated with a RLock because
the GC doesn't actually stop other threads -- other threads can acquire
the GIL and run during many parts of a GC cycle.

For ease of implementation, we add a private API _thread.CriticalLock
which disables the current thread from parking while the lock is held.
This prevents deadlock by ensuring that the lock is released before
the GC is run.